feat/refactor : Initialize Presentation Module + Initial Room Database impl
Since the very beginning for this app development, we're merged the presentation layer in the app module. And also in this app module,
we're implementing the Dependency Injection here (we're using Koin) since Koin needs the Android Context which only app module can provide.

But, the problem is, we're using Dependency Injection for all layers in this app (data, domain, presentation), so this app module know every implementation from all of those modules because it have to.
And because of that, it's kind of breaking the Clean Architecture principle since now Presentation layer (that merged into the app module) know everything about Data layers implementation, which it shouldn't.

So, i have to refactor the project to separate this presentation layer into it's own module and the app module right now act as a "glue" for all the layers to the Android Context, which is turns out to
be fine and make the Separation of Concern even more Cleaner and doesn't break the Clean Architecture Implementations.

Big Thanks to these references :
https://github.com/InsertKoinIO/koin/issues/472
https://github.com/tamimattafi/just-do-it